Move control sizer into its own file
